feat: otherworldly performance enhancements reducing cloning

+++ server/src/game/square.rs
@@ -9,14 +9,14 @@ pub fn square_to_vector(square: &str) -> Vector {
     let rank = chars.next().unwrap();
     let file = file as u8 - 'a' as u8;
     let rank = rank as u8 - '1' as u8;
-    Vector::new(rank as i32, file as i32)
+    Vector::new(rank, file)
 }
 
 impl Display for PossibleMove {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let mut result = String::new();
-        let from = format!("{}{}", (self.origin.x + 1) as u8 as char, (self.origin.y + 1) as u8 as char);
-        let to = format!("{}{}", (self.target.x + 1) as u8 as char, (self.target.y + 1) as u8 as char);
+        let from = format!("{}{}", (self.origin.x + 1) as char, (self.origin.y + 1) as char);
+        let to = format!("{}{}", (self.target.x + 1) as char, (self.target.y + 1) as char);
         result.push_str(&from);
         result.push_str(&to);
         write!(f, "{}", result)
